requirements.txt
selenium==4.9.1
webdriver-manager
python-dotenv==1.0.0
spacy==3.5.2
openai==0.27.8
skillNer==1.0.3
nltk==3.8.1
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.5.0/en_core_web_lg-3.5.0-py3-none-any.whl#sha256=c8ac64840c1eb3e3ca7bd38bd1e1c48fb0faeb2449d54d01d5ce629af4595775
NUGigSkillNER==1.0.2